+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Juniper.jl Changelog
 
+### v0.1.2
+- Bugfix: Reset of `mu_init` in Ipopt options to have the default `mu_init` if `solve` is called again
+- Bugfix: Break on time limit in relaxation, fpump and strong branching
+
 ### v0.1.1
 - Freemodel for commerical nlp solvers with license restrictions
 - More convenient parallel options